backend/pyproject.toml
[project]
name = "practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "numpy>=2.4.1",
    "openai>=2.16.0",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "twilio>=9.10.0",
    "uvicorn>=0.40.0",
]
